docker_stuff/requirements.txt
#Stuff here yet!

sqlalchemy
psycopg2-binary
pandas
streamlit